Spectrum: Allow kernel developers to remove support for specific profiles
Not all kernel developers ship all 4 profiles Spectrum supports, so add a
simple mechanism for kernel devs to be able to get rid of no-op tiles.

set the spectrum.disabledprofiles prop as a comma separated list of profiles
you DO NOT ship to remove the card for it from the Spectrum UI. Valid profile
names are balance, performance, gaming, and battery.